Use float literals

+++ src/raytracer/vector3d.rs
@@ -90,7 +90,7 @@ pub fn unit_vector(v: &Vector3d) -> Vector3d {
 #[inline(always)]
 pub fn random_unit_vector(rng: &mut StdRng) -> Vector3d {
     let a = rng.gen_range(0.0, 2.0 * PI);
-    let z = rng.gen_range(-1.0 as f64, 1.0 as f64);
+    let z = rng.gen_range(-1.0_f64, 1.0_f64);
     let r = (1.0 - z * z).sqrt();
     Vector3d {
         x: r * a.cos(),
@@ -101,7 +101,7 @@ pub fn random_unit_vector(rng: &mut StdRng) -> Vector3d {
 
 #[inline(always)]
 pub fn reflect(v: &Vector3d, n: &Vector3d) -> Vector3d {
-    *v - &((*n * dot(v, n)) * 2.0 as f64)
+    *v - &((*n * dot(v, n)) * 2.0_f64)
 }
 
 #[inline(always)]
